+++ core/bin/core_api.rs
@@ -28,7 +28,8 @@ use tokio_stream::Stream;
 use tracing::{error, info};
 
 use dust::api::api_state::APIState;
-use dust::api::projects;
+use dust::api::data_sources::DataSourcesDocumentsUpdateParentsPayload;
+use dust::api::{data_sources, datasets, projects, runs, specifications};
 use dust::{
     api_keys::validate_api_key,
     app,
@@ -70,2260 +71,6 @@ static GLOBAL: Jemalloc = Jemalloc;
 
 async fn index() -> &'static str {
     "dust_api server ready"
-}
-
-/// Check a specification
-
-#[derive(serde::Deserialize)]
-struct SpecificationsCheckPayload {
-    specification: String,
-}
-
-async fn specifications_check(
-    Path(project_id): Path<i64>,
-    Json(payload): Json<SpecificationsCheckPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let _project = project::Project::new_from_id(project_id);
-    match app::App::new(&payload.specification).await {
-        Err(e) => error_response(
-            StatusCode::BAD_REQUEST,
-            "invalid_specification_error",
-            "Invalid specification",
-            Some(e),
-        ),
-        Ok(app) => (
-            StatusCode::OK,
-            Json(APIResponse {
-                error: None,
-                response: Some(json!({
-                    "app": {
-                        "hash": app.hash(),
-                        "blocks": app.blocks(),
-                    }
-                })),
-            }),
-        ),
-    }
-}
-
-/// Retrieve a specification
-
-async fn specifications_retrieve(
-    Path((project_id, hash)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state.store.load_specification(&project, &hash).await {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve specification",
-            Some(e),
-        ),
-        Ok(s) => match s {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "specification_not_found",
-                &format!("No specification found with hash `{}`", hash),
-                None,
-            ),
-            Some((created, spec)) => (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "specification": {
-                            "created": created,
-                            "data": spec,
-                        },
-                    })),
-                }),
-            ),
-        },
-    }
-}
-
-async fn specifications_get(
-    Path(project_id): Path<i64>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-
-    match state.store.list_specification_hashes(&project).await {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to list specifications",
-            Some(e),
-        ),
-        Ok(hashes) => (
-            StatusCode::OK,
-            Json(APIResponse {
-                error: None,
-                response: Some(json!({ "hashes": hashes })),
-            }),
-        ),
-    }
-}
-
-/// Push a specification
-
-#[derive(serde::Deserialize)]
-struct SpecificationsPushPayload {
-    specification: String,
-}
-
-async fn specifications_post(
-    Path(project_id): Path<i64>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<SpecificationsPushPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    match save_specification(project_id, &state.store, payload.specification).await {
-        Err(err) => err,
-        Ok(app) => (
-            StatusCode::OK,
-            Json(APIResponse {
-                error: None,
-                response: Some(json!({
-                    "app":{
-                        "hash": app.hash()
-                     },
-                })),
-            }),
-        ),
-    }
-}
-
-async fn save_specification(
-    project_id: i64,
-    store: &Box<dyn store::Store + Sync + Send>,
-    specification: String,
-) -> Result<app::App, (StatusCode, Json<APIResponse>)> {
-    let project = project::Project::new_from_id(project_id);
-
-    let app = match app::App::new(&specification).await {
-        Err(e) => Err(error_response(
-            StatusCode::BAD_REQUEST,
-            "invalid_specification_error",
-            "Invalid specification",
-            Some(e),
-        ))?,
-        Ok(app) => app,
-    };
-
-    match store
-        .register_specification(&project, &app.hash(), &specification)
-        .await
-    {
-        Err(e) => Err(error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to register specification",
-            Some(e),
-        ))?,
-        Ok(_) => (),
-    }
-    Ok(app)
-}
-
-/// Register a new dataset
-
-#[derive(serde::Deserialize)]
-struct DatasetsRegisterPayload {
-    dataset_id: String,
-    data: Vec<Value>,
-}
-
-async fn datasets_register(
-    Path(project_id): Path<i64>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<DatasetsRegisterPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match dataset::Dataset::new_from_jsonl(&payload.dataset_id, payload.data).await {
-        Err(e) => error_response(
-            StatusCode::BAD_REQUEST,
-            "invalid_dataset_error",
-            "Invalid dataset",
-            Some(e),
-        ),
-        Ok(d) => {
-            // First retrieve the latest hash of the dataset to avoid registering if it matches the
-            // current hash.
-            let current_hash = state
-                .store
-                .latest_dataset_hash(&project, &d.dataset_id())
-                .await
-                .unwrap_or_else(|_| None);
-            if !(current_hash.is_some() && current_hash.unwrap() == d.hash()) {
-                match state.store.register_dataset(&project, &d).await {
-                    Err(e) => error_response(
-                        StatusCode::INTERNAL_SERVER_ERROR,
-                        "internal_server_error",
-                        "Failed to store dataset",
-                        Some(e),
-                    ),
-                    Ok(()) => (
-                        StatusCode::OK,
-                        Json(APIResponse {
-                            error: None,
-                            response: Some(json!({
-                                "dataset": {
-                                    "created": d.created(),
-                                    "dataset_id": d.dataset_id(),
-                                    "hash": d.hash(),
-                                    "keys": d.keys(),
-                                }
-                            })),
-                        }),
-                    ),
-                }
-            } else {
-                (
-                    StatusCode::OK,
-                    Json(APIResponse {
-                        error: None,
-                        response: Some(json!({
-                            "dataset": {
-                                "created": d.created(),
-                                "dataset_id": d.dataset_id(),
-                                "hash": d.hash(),
-                                "keys": d.keys(),
-                            }
-                        })),
-                    }),
-                )
-            }
-        }
-    }
-}
-
-#[tracing::instrument(level = "info", skip_all)]
-async fn datasets_list(
-    Path(project_id): Path<i64>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-
-    match state.store.list_datasets(&project).await {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to list datasets",
-            Some(e),
-        ),
-        Ok(datasets) => {
-            let datasets = datasets
-                .into_iter()
-                .map(|(d, v)| {
-                    (
-                        d,
-                        v.into_iter()
-                            .map(|(h, c)| {
-                                json!({
-                                    "hash": h,
-                                    "created": c,
-                                })
-                            })
-                            .collect::<Vec<_>>(),
-                    )
-                })
-                .collect::<HashMap<_, _>>();
-            (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "datasets": datasets,
-                    })),
-                }),
-            )
-        }
-    }
-}
-
-async fn datasets_retrieve(
-    Path((project_id, dataset_id, hash)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state.store.load_dataset(&project, &dataset_id, &hash).await {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve dataset",
-            Some(e),
-        ),
-        Ok(dataset) => match dataset {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "dataset_not_found",
-                &format!(
-                    "No dataset found for id `{}` and hash `{}`",
-                    dataset_id, hash,
-                ),
-                None,
-            ),
-            Some(d) => (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "dataset": d,
-                    })),
-                }),
-            ),
-        },
-    }
-}
-
-#[derive(Clone, serde::Deserialize)]
-struct Secret {
-    name: String,
-    value: String,
-}
-
-#[derive(serde::Deserialize, Clone)]
-struct RunsCreatePayload {
-    run_type: run::RunType,
-    specification: Option<String>,
-    specification_hash: Option<String>,
-    dataset_id: Option<String>,
-    inputs: Option<Vec<Value>>,
-    config: run::RunConfig,
-    credentials: run::Credentials,
-    secrets: Vec<Secret>,
-    store_blocks_results: Option<bool>,
-}
-
-async fn run_helper(
-    project_id: i64,
-    payload: RunsCreatePayload,
-    state: Arc<APIState>,
-) -> Result<app::App, (StatusCode, Json<APIResponse>)> {
-    let project = project::Project::new_from_id(project_id);
-
-    let mut register_spec = true;
-    let specification = match payload.specification {
-        Some(spec) => spec,
-        None => match payload.specification_hash {
-            Some(hash) => {
-                let hash = match hash.as_str() {
-                    "latest" => match state.store.latest_specification_hash(&project).await {
-                        Err(e) => Err(error_response(
-                            StatusCode::INTERNAL_SERVER_ERROR,
-                            "internal_server_error",
-                            "Failed to retrieve latest specification",
-                            Some(e),
-                        ))?,
-                        Ok(h) => match h {
-                            None => Err(error_response(
-                                StatusCode::NOT_FOUND,
-                                "specification_not_found",
-                                "Latest specification not found",
-                                None,
-                            ))?,
-                            Some(h) => h,
-                        },
-                    },
-                    _ => hash,
-                };
-
-                match state.store.load_specification(&project, &hash).await {
-                    Err(e) => Err(error_response(
-                        StatusCode::INTERNAL_SERVER_ERROR,
-                        "internal_server_error",
-                        "Failed to retrieve specification",
-                        Some(e),
-                    ))?,
-                    Ok(spec) => match spec {
-                        None => Err(error_response(
-                            StatusCode::NOT_FOUND,
-                            "specification_not_found",
-                            &format!("No specification found for hash `{}`", hash),
-                            None,
-                        ))?,
-                        Some((_, s)) => {
-                            register_spec = false;
-                            s
-                        }
-                    },
-                }
-            }
-            None => Err(error_response(
-                StatusCode::BAD_REQUEST,
-                "missing_specification_error",
-                "No specification provided, either `specification` \
-                 or `specification_hash` must be provided",
-                None,
-            ))?,
-        },
-    };
-
-    let mut app = match app::App::new(&specification).await {
-        Err(e) => Err(error_response(
-            StatusCode::BAD_REQUEST,
-            "invalid_specification_error",
-            "Invalid specification",
-            Some(e),
-        ))?,
-        Ok(app) => app,
-    };
-
-    let mut d = match payload.dataset_id.as_ref() {
-        None => None,
-        Some(dataset_id) => match state.store.latest_dataset_hash(&project, dataset_id).await {
-            Err(e) => Err(error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "internal_server_error",
-                "Failed to retrieve dataset",
-                Some(e),
-            ))?,
-            Ok(None) => Err(error_response(
-                StatusCode::NOT_FOUND,
-                "dataset_not_found",
-                &format!("No dataset found for id `{}`", dataset_id),
-                None,
-            ))?,
-            Ok(Some(latest)) => match state
-                .store
-                .load_dataset(&project, dataset_id, &latest)
-                .await
-            {
-                Err(e) => Err(error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to retrieve dataset",
-                    Some(e),
-                ))?,
-                Ok(d) => match d {
-                    None => unreachable!(),
-                    Some(d) => Some(d),
-                },
-            },
-        },
-    };
-
-    if d.is_some() {
-        if payload.run_type != run::RunType::Local {
-            Err(error_response(
-                StatusCode::BAD_REQUEST,
-                "invalid_run_type_error",
-                "RunType `local` is expected when a `dataset_id` is provided",
-                None,
-            ))?
-        }
-
-        if d.as_ref().unwrap().len() == 0 {
-            Err(error_response(
-                StatusCode::BAD_REQUEST,
-                "dataset_empty_error",
-                &format!(
-                    "Dataset `{}` has 0 record",
-                    payload.dataset_id.as_ref().unwrap()
-                ),
-                None,
-            ))?
-        }
-
-        info!(
-            dataset_id = payload.dataset_id.as_ref().unwrap(),
-            records = d.as_ref().unwrap().len(),
-            "Retrieved latest version of dataset"
-        );
-    }
-
-    if payload.inputs.is_some() {
-        d = match dataset::Dataset::new_from_jsonl("inputs", payload.inputs.unwrap()).await {
-            Err(e) => Err(error_response(
-                StatusCode::BAD_REQUEST,
-                "invalid_inputs_error",
-                "Invalid inputs",
-                Some(e),
-            ))?,
-            Ok(d) => Some(d),
-        };
-        info!(records = d.as_ref().unwrap().len(), "Received inputs");
-    }
-
-    // Only register the specification if it was not passed by hash.
-    if register_spec {
-        match state
-            .store
-            .register_specification(&project, &app.hash(), &specification)
-            .await
-        {
-            Err(e) => Err(error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "internal_server_error",
-                "Failed to register specification",
-                Some(e),
-            ))?,
-            Ok(_) => (),
-        }
-    }
-
-    match app
-        .prepare_run(
-            payload.run_type,
-            payload.config,
-            project.clone(),
-            d,
-            state.store.clone(),
-        )
-        .await
-    {
-        Err(e) => Err(error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed prepare run",
-            Some(e),
-        ))?,
-        Ok(()) => (),
-    }
-
-    Ok(app)
-}
-
-async fn runs_create(
-    Path(project_id): Path<i64>,
-    headers: HeaderMap,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<RunsCreatePayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let mut credentials = payload.credentials.clone();
-
-    // Convert payload secrets vector to hash map to use them with {secrets.SECRET_NAME}.
-    let secrets = run::Secrets {
-        redacted: true,
-        secrets: payload
-            .secrets
-            .iter()
-            .map(|secret| (secret.name.clone(), secret.value.clone()))
-            .collect::<HashMap<_, _>>(),
-    };
-
-    match headers.get("X-Dust-Workspace-Id") {
-        Some(v) => match v.to_str() {
-            Ok(v) => {
-                credentials.insert("DUST_WORKSPACE_ID".to_string(), v.to_string());
-            }
-            _ => (),
-        },
-        None => (),
-    };
-
-    match headers.get("X-Dust-Feature-Flags") {
-        Some(v) => match v.to_str() {
-            Ok(v) => {
-                credentials.insert("DUST_FEATURE_FLAGS".to_string(), v.to_string());
-            }
-            _ => (),
-        },
-        None => (),
-    };
-
-    match headers.get("X-Dust-Group-Ids") {
-        Some(v) => match v.to_str() {
-            Ok(v) => {
-                credentials.insert("DUST_GROUP_IDS".to_string(), v.to_string());
-            }
-            _ => (),
-        },
-        None => (),
-    };
-
-    // If the run is made by a system key, it's a system run
-    match headers.get("X-Dust-IsSystemRun") {
-        Some(v) => match v.to_str() {
-            Ok(v) => {
-                credentials.insert("DUST_IS_SYSTEM_RUN".to_string(), v.to_string());
-            }
-            _ => (),
-        },
-        None => (),
-    };
-
-    match run_helper(project_id, payload.clone(), state.clone()).await {
-        Ok(app) => {
-            // The run is empty for now, we can clone it for the response.
-            let run = app.run_ref().unwrap().clone();
-            state.run_app(
-                app,
-                credentials,
-                secrets,
-                payload.store_blocks_results.unwrap_or(true),
-            );
-            (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "run": run,
-                    })),
-                }),
-            )
-        }
-        Err(err) => err,
-    }
-}
-
-async fn runs_create_stream(
-    Path(project_id): Path<i64>,
-    headers: HeaderMap,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<RunsCreatePayload>,
-) -> Sse<impl Stream<Item = Result<Event, Infallible>>> {
-    let mut credentials = payload.credentials.clone();
-
-    // Convert payload secrets vector to hash map to use them with {secrets.SECRET_NAME}.
-    let secrets = run::Secrets {
-        redacted: true,
-        secrets: payload
-            .secrets
-            .iter()
-            .map(|secret| (secret.name.clone(), secret.value.clone()))
-            .collect::<HashMap<_, _>>(),
-    };
-
-    match headers.get("X-Dust-Workspace-Id") {
-        Some(v) => match v.to_str() {
-            Ok(v) => {
-                credentials.insert("DUST_WORKSPACE_ID".to_string(), v.to_string());
-            }
-            _ => (),
-        },
-        None => (),
-    };
-
-    match headers.get("X-Dust-Feature-Flags") {
-        Some(v) => match v.to_str() {
-            Ok(v) => {
-                credentials.insert("DUST_FEATURE_FLAGS".to_string(), v.to_string());
-            }
-            _ => (),
-        },
-        None => (),
-    };
-
-    match headers.get("X-Dust-Group-Ids") {
-        Some(v) => match v.to_str() {
-            Ok(v) => {
-                credentials.insert("DUST_GROUP_IDS".to_string(), v.to_string());
-            }
-            _ => (),
-        },
-        None => (),
-    };
-
-    // If the run is made by a system key, it's a system run
-    match headers.get("X-Dust-IsSystemRun") {
-        Some(v) => match v.to_str() {
-            Ok(v) => {
-                credentials.insert("DUST_IS_SYSTEM_RUN".to_string(), v.to_string());
-            }
-            _ => (),
-        },
-        None => (),
-    };
-
-    // create unbounded channel to pass as stream to Sse::new
-    let (tx, mut rx) = unbounded_channel::<Value>();
-
-    match run_helper(project_id, payload.clone(), state.clone()).await {
-        Ok(mut app) => {
-            // The run is empty for now, we can clone it for the response.
-            // let run = app.run_ref().unwrap().clone();
-            let store = state.store.clone();
-            let databases_store = state.databases_store.clone();
-            let qdrant_clients = state.qdrant_clients.clone();
-
-            // Start a task that will run the app in the background.
-            tokio::task::spawn(async move {
-                let now = std::time::Instant::now();
-                match app
-                    .run(
-                        credentials,
-                        secrets,
-                        store,
-                        databases_store,
-                        qdrant_clients,
-                        Some(tx.clone()),
-                        payload.store_blocks_results.unwrap_or(true),
-                    )
-                    .await
-                {
-                    Ok(()) => {
-                        info!(
-                            run = app.run_ref().unwrap().run_id(),
-                            app_version = app.hash(),
-                            elapsed = now.elapsed().as_millis(),
-                            "Run finished"
-                        );
-                    }
-                    Err(e) => {
-                        error!(error = %e, "Run error");
-                    }
-                }
-            });
-        }
-        Err((_, api_error)) => {
-            let error = api_error.0.error.unwrap_or_else(|| APIError {
-                code: "internal_server_error".to_string(),
-                message: "The app execution failed unexpectedly".to_string(),
-            });
-            let _ = tx.send(json!({
-                "type": "error",
-                "content": {
-                    "code": error.code,
-                    "message": error.message,
-                },
-            }));
-        }
-    }
-
-    let stream = async_stream::stream! {
-        while let Some(v) = rx.recv().await {
-            match Event::default().json_data(v) {
-                Ok(event) => yield Ok(event),
-                Err(e) => {
-                    error!(error = %e, "Failed to create SSE event");
-                }
-            };
-        }
-        match Event::default().json_data(json!({
-            "type": "final",
-            "content": null,
-        })) {
-            Ok(event) => yield Ok(event),
-            Err(e) => {
-                error!(error = %e, "Failed to create SSE event");
-            }
-        };
-    };
-
-    Sse::new(stream).keep_alive(KeepAlive::default())
-}
-
-async fn runs_delete(
-    Path((project_id, run_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state.store.delete_run(&project, &run_id).await {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to delete run",
-            Some(e),
-        ),
-        Ok(()) => (
-            StatusCode::OK,
-            Json(APIResponse {
-                error: None,
-                response: Some(json!({
-                    "success": true
-                })),
-            }),
-        ),
-    }
-}
-
-async fn runs_cancel(
-    Path((project_id, run_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-
-    // Load the run and update its status to cancelled (errored)
-    match state.store.load_run(&project, &run_id, None).await {
-        Ok(Some(mut run)) => {
-            // Check if the run is at least 1 hour old before allowing cancellation
-            let current_time = utils::now();
-            let run_age_ms = current_time - run.created();
-            const ONE_HOUR_MS: u64 = 60 * 60 * 1000; // 1 hour in milliseconds
-
-            if run_age_ms < ONE_HOUR_MS {
-                info!(
-                    run = run_id,
-                    age_ms = run_age_ms,
-                    "Run is too recent to cancel (must be at least 1 hour old)"
-                );
-                return error_response(
-                    StatusCode::BAD_REQUEST,
-                    "run_too_recent",
-                    "Run must be at least 1 hour old before it can be cancelled",
-                    None,
-                );
-            }
-
-            // Cancel the run (marks run and all running blocks as errored)
-            run.cancel();
-
-            // Update the run status in the database
-            match state
-                .store
-                .update_run_status(&project, &run_id, run.status())
-                .await
-            {
-                Err(e) => {
-                    error!(error = %e, run = run_id, "Failed to update run status to cancelled");
-                    error_response(
-                        StatusCode::INTERNAL_SERVER_ERROR,
-                        "internal_server_error",
-                        "Failed to cancel run",
-                        Some(e),
-                    )
-                }
-                Ok(_) => {
-                    info!(run = run_id, "Run cancelled successfully");
-                    (
-                        StatusCode::OK,
-                        Json(APIResponse {
-                            error: None,
-                            response: Some(json!({
-                                "success": true
-                            })),
-                        }),
-                    )
-                }
-            }
-        }
-        Ok(None) => {
-            // Run not found in database
-            info!(run = run_id, "Run not found for cancellation");
-            (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "success": true
-                    })),
-                }),
-            )
-        }
-        Err(e) => {
-            error!(error = %e, run = run_id, "Failed to load run for cancellation");
-            error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "internal_server_error",
-                "Failed to load run",
-                Some(e),
-            )
-        }
-    }
-}
-
-#[derive(serde::Deserialize)]
-struct RunsListQuery {
-    offset: usize,
-    limit: usize,
-    run_type: run::RunType,
-}
-
-async fn runs_list(
-    Path(project_id): Path<i64>,
-    State(state): State<Arc<APIState>>,
-    Query(query): Query<RunsListQuery>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .list_runs(&project, query.run_type, Some((query.limit, query.offset)))
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to list runs",
-            Some(e),
-        ),
-        Ok((runs, total)) => (
-            StatusCode::OK,
-            Json(APIResponse {
-                error: None,
-                response: Some(json!({
-                    "offset": query.offset,
-                    "limit": query.limit,
-                    "total": total,
-                    "runs": runs,
-                })),
-            }),
-        ),
-    }
-}
-
-#[derive(serde::Deserialize)]
-struct RunsRetrieveBatchPayload {
-    run_ids: Vec<String>,
-}
-
-async fn runs_retrieve_batch(
-    Path(project_id): Path<i64>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<RunsRetrieveBatchPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state.store.load_runs(&project, payload.run_ids).await {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve runs",
-            Some(e),
-        ),
-        Ok(runs) => (
-            StatusCode::OK,
-            Json(APIResponse {
-                error: None,
-                response: Some(json!({
-                    "runs": runs,
-                })),
-            }),
-        ),
-    }
-}
-
-async fn runs_retrieve(
-    Path((project_id, run_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state.store.load_run(&project, &run_id, None).await {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve run",
-            Some(e),
-        ),
-        Ok(run) => match run {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "run_not_found",
-                &format!("No run found for id `{}`", run_id),
-                None,
-            ),
-            Some(run) => (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "run": run,
-                    })),
-                }),
-            ),
-        },
-    }
-}
-
-async fn runs_retrieve_block(
-    Path((project_id, run_id, block_type, block_name)): Path<(i64, String, BlockType, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .load_run(&project, &run_id, Some(Some((block_type, block_name))))
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve run",
-            Some(e),
-        ),
-        Ok(run) => match run {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "run_not_found",
-                &format!("No run found for id `{}`", run_id),
-                None,
-            ),
-            Some(run) => (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "run": run,
-                    })),
-                }),
-            ),
-        },
-    }
-}
-
-async fn runs_retrieve_status(
-    Path((project_id, run_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state.store.load_run(&project, &run_id, Some(None)).await {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve run",
-            Some(e),
-        ),
-        Ok(run) => match run {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "run_not_found",
-                &format!("No run found for id `{}`", run_id),
-                None,
-            ),
-            Some(run) => (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "run": run,
-                    })),
-                }),
-            ),
-        },
-    }
-}
-
-/// Register a new data source.
-
-#[derive(serde::Deserialize)]
-struct DataSourcesRegisterPayload {
-    config: data_source::DataSourceConfig,
-    #[allow(dead_code)]
-    credentials: run::Credentials,
-    name: String,
-}
-
-async fn data_sources_register(
-    Path(project_id): Path<i64>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<DataSourcesRegisterPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    let ds = data_source::DataSource::new(&project, &payload.config, &payload.name);
-
-    match ds
-        .register(state.store.clone(), state.search_store.clone())
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to register data source",
-            Some(e),
-        ),
-        Ok(()) => (
-            StatusCode::OK,
-            Json(APIResponse {
-                error: None,
-                response: Some(json!({
-                    "data_source": {
-                        "created": ds.created(),
-                        "data_source_id": ds.data_source_id(),
-                        "name": ds.name(),
-                        "config": ds.config(),
-                    },
-                })),
-            }),
-        ),
-    }
-}
-
-/// Update a data source.
-
-#[derive(serde::Deserialize)]
-struct DataSourcesUpdatePayload {
-    name: String,
-}
-
-async fn data_sources_update(
-    Path((project_id, data_source_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<DataSourcesUpdatePayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-
-    let mut ds = match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => {
-            return error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "internal_server_error",
-                "Failed to retrieve data source",
-                Some(e),
-            );
-        }
-        Ok(None) => {
-            return error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            );
-        }
-        Ok(Some(ds)) => ds,
-    };
-
-    if let Err(e) = ds
-        .update_name(
-            state.store.clone(),
-            state.search_store.clone(),
-            &payload.name,
-        )
-        .await
-    {
-        return error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to update data source name",
-            Some(e),
-        );
-    }
-
-    (
-        StatusCode::OK,
-        Json(APIResponse {
-            error: None,
-            response: Some(json!({
-                "data_source": {
-                    "created": ds.created(),
-                    "data_source_id": ds.data_source_id(),
-                    "name": ds.name(),
-                    "config": ds.config(),
-                },
-            })),
-        }),
-    )
-}
-
-#[derive(serde::Deserialize)]
-struct DataSourcesTokenizePayload {
-    text: String,
-}
-async fn data_sources_tokenize(
-    Path((project_id, data_source_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<DataSourcesTokenizePayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => {
-            error!(
-                error = %e,
-                project_id = project_id,
-                data_source_id = %data_source_id,
-                "Failed to retrieve data source"
-            );
-            error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "internal_server_error",
-                "Failed to retrieve data source",
-                Some(e),
-            )
-        }
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => {
-                let embedder_config = ds.embedder_config().clone();
-                let provider_id = embedder_config.provider_id;
-                let model_id = embedder_config.model_id.clone();
-                let embedder =
-                    provider(embedder_config.provider_id).embedder(embedder_config.model_id);
-                match embedder.tokenize(vec![payload.text]).await {
-                    Err(e) => {
-                        error!(
-                            error = %e,
-                            project_id = project_id,
-                            data_source_id = %data_source_id,
-                            provider_id = %provider_id,
-                            model_id = %model_id,
-                            "Failed to tokenize text"
-                        );
-                        error_response(
-                            StatusCode::INTERNAL_SERVER_ERROR,
-                            "internal_server_error",
-                            "Failed to tokenize text",
-                            Some(e),
-                        )
-                    }
-                    Ok(mut res) => match res.pop() {
-                        None => {
-                            error!(
-                                project_id = project_id,
-                                data_source_id = %data_source_id,
-                                "Tokenizer returned empty result"
-                            );
-                            error_response(
-                                StatusCode::INTERNAL_SERVER_ERROR,
-                                "internal_server_error",
-                                "Failed to tokenize text",
-                                None,
-                            )
-                        }
-                        Some(tokens) => (
-                            StatusCode::OK,
-                            Json(APIResponse {
-                                error: None,
-                                response: Some(json!({
-                                    "tokens": tokens,
-                                })),
-                            }),
-                        ),
-                    },
-                }
-            }
-        },
-    }
-}
-
-async fn data_sources_retrieve(
-    Path((project_id, data_source_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "data_source": {
-                            "created": ds.created(),
-                            "data_source_id": ds.data_source_id(),
-                            "data_source_internal_id": ds.internal_id(),
-                            "config": ds.config(),
-                            "name": ds.name(),
-                        },
-                    })),
-                }),
-            ),
-        },
-    }
-}
-
-// Perform a search on a data source.
-
-#[derive(serde::Deserialize)]
-struct DatasourceSearchPayload {
-    query: Option<String>,
-    top_k: usize,
-    filter: Option<SearchFilter>,
-    view_filter: Option<SearchFilter>,
-    full_text: bool,
-    credentials: run::Credentials,
-    target_document_tokens: Option<usize>,
-}
-
-async fn data_sources_search(
-    Path((project_id, data_source_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<DatasourceSearchPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => match ds
-                .search(
-                    payload.credentials,
-                    state.store.clone(),
-                    state.qdrant_clients.clone(),
-                    &payload.query,
-                    payload.top_k,
-                    match payload.filter {
-                        Some(filter) => Some(filter.postprocess_for_data_source(&data_source_id)),
-                        None => None,
-                    },
-                    match payload.view_filter {
-                        Some(filter) => Some(filter.postprocess_for_data_source(&data_source_id)),
-                        None => None,
-                    },
-                    payload.full_text,
-                    payload.target_document_tokens,
-                )
-                .await
-            {
-                Ok(documents) => (
-                    StatusCode::OK,
-                    Json(APIResponse {
-                        error: None,
-                        response: Some(json!({
-                            "documents": documents,
-                        })),
-                    }),
-                ),
-                Err(e) => error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to perform the search",
-                    Some(e),
-                ),
-            },
-        },
-    }
-}
-
-/// Update tags of a document in a data source.
-
-#[derive(serde::Deserialize)]
-struct DataSourcesDocumentsUpdateTagsPayload {
-    add_tags: Option<Vec<String>>,
-    remove_tags: Option<Vec<String>>,
-}
-
-async fn data_sources_documents_update_tags(
-    Path((project_id, data_source_id, document_id)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<DataSourcesDocumentsUpdateTagsPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    let add_tags = payload.add_tags.unwrap_or_else(|| vec![]);
-    let remove_tags = payload.remove_tags.unwrap_or_else(|| vec![]);
-    let add_tags_set: HashSet<String> = add_tags.iter().cloned().collect();
-    let remove_tags_set: HashSet<String> = remove_tags.iter().cloned().collect();
-
-    if add_tags_set.intersection(&remove_tags_set).count() > 0 {
-        return error_response(
-            StatusCode::BAD_REQUEST,
-            "bad_request",
-            "The `add_tags` and `remove_tags` lists have a non-empty intersection.",
-            None,
-        );
-    }
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => match ds
-                .update_tags(
-                    state.store.clone(),
-                    state.qdrant_clients.clone(),
-                    document_id,
-                    add_tags_set.into_iter().collect(),
-                    remove_tags_set.into_iter().collect(),
-                )
-                .await
-            {
-                Err(e) => error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to update document tags",
-                    Some(e),
-                ),
-                Ok(_) => (
-                    StatusCode::OK,
-                    Json(APIResponse {
-                        error: None,
-                        response: Some(json!({
-                            "data_source": {
-                                "created": ds.created(),
-                                "data_source_id": ds.data_source_id(),
-                                "config": ds.config(),
-                            },
-                        })),
-                    }),
-                ),
-            },
-        },
-    }
-}
-
-/// Update parents of a document in a data source.
-
-#[derive(serde::Deserialize)]
-struct DataSourcesDocumentsUpdateParentsPayload {
-    parent_id: Option<String>,
-    parents: Vec<String>,
-}
-
-async fn data_sources_documents_update_parents(
-    Path((project_id, data_source_id, document_id)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<DataSourcesDocumentsUpdateParentsPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-
-    if payload.parents.get(0) != Some(&document_id) {
-        return error_response(
-            StatusCode::BAD_REQUEST,
-            "invalid_parents",
-            "Failed to update document parents - parents[0] and document_id should be equal",
-            None,
-        );
-    }
-
-    match &payload.parent_id {
-        Some(parent_id) => {
-            if payload.parents.get(1) != Some(parent_id) {
-                return error_response(
-                    StatusCode::BAD_REQUEST,
-                    "invalid_parent_id",
-                    "Failed to update document parents - parents[1] and parent_id should be equal",
-                    None,
-                );
-            }
-        }
-        None => {
-            if payload.parents.len() > 1 {
-                return error_response(
-                    StatusCode::BAD_REQUEST,
-                    "invalid_parent_id",
-                    "Failed to update document parents - parent_id should not be null if parents[1] is defined",
-                    None,
-                );
-            }
-        }
-    }
-
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => match ds
-                .update_parents(
-                    state.store.clone(),
-                    state.qdrant_clients.clone(),
-                    state.search_store.clone(),
-                    document_id,
-                    payload.parents,
-                )
-                .await
-            {
-                Err(e) => error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to update document parents",
-                    Some(e),
-                ),
-                Ok(_) => (
-                    StatusCode::OK,
-                    Json(APIResponse {
-                        error: None,
-                        response: Some(json!({
-                            "data_source": {
-                                "created": ds.created(),
-                                "data_source_id": ds.data_source_id(),
-                                "config": ds.config(),
-                            },
-                        })),
-                    }),
-                ),
-            },
-        },
-    }
-}
-
-// List versions of a document in a data source.
-
-#[derive(serde::Deserialize)]
-struct DataSourcesDocumentsVersionsListQuery {
-    offset: usize,
-    limit: usize,
-    latest_hash: Option<String>, // Hash of the latest version to retrieve.
-    view_filter: Option<String>, // Parsed as JSON.
-}
-
-async fn data_sources_documents_versions_list(
-    Path((project_id, data_source_id, document_id)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-    Query(query): Query<DataSourcesDocumentsVersionsListQuery>,
-) -> (StatusCode, Json<APIResponse>) {
-    let view_filter: Option<SearchFilter> = match query
-        .view_filter
-        .as_ref()
-        .and_then(|f| Some(serde_json::from_str(f)))
-    {
-        Some(Ok(f)) => Some(f),
-        None => None,
-        Some(Err(e)) => {
-            return error_response(
-                StatusCode::BAD_REQUEST,
-                "invalid_view_filter",
-                "Failed to parse view_filter query parameter",
-                Some(e.into()),
-            )
-        }
-    };
-
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .list_data_source_document_versions(
-            &project,
-            &data_source_id,
-            &document_id,
-            Some((query.limit, query.offset)),
-            &match view_filter {
-                Some(filter) => Some(filter.postprocess_for_data_source(&data_source_id)),
-                None => None,
-            },
-            &query.latest_hash,
-            true,
-        )
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to list document versions",
-            Some(e),
-        ),
-        Ok((versions, total)) => (
-            StatusCode::OK,
-            Json(APIResponse {
-                error: None,
-                response: Some(json!({
-                    "offset": query.offset,
-                    "limit": query.limit,
-                    "total": total,
-                    "versions": versions,
-                })),
-            }),
-        ),
-    }
-}
-
-/// Upsert a document in a data source.
-
-#[derive(serde::Deserialize)]
-struct DataSourcesDocumentsUpsertPayload {
-    document_id: String,
-    timestamp: Option<u64>,
-    tags: Vec<String>,
-    parent_id: Option<String>,
-    parents: Vec<String>,
-    source_url: Option<String>,
-    section: Section,
-    credentials: run::Credentials,
-    light_document_output: Option<bool>,
-    title: String,
-    mime_type: String,
-    provider_visibility: Option<ProviderVisibility>,
-}
-
-async fn data_sources_documents_upsert(
-    Path((project_id, data_source_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-    Json(payload): Json<DataSourcesDocumentsUpsertPayload>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    let light_document_output = payload.light_document_output.unwrap_or_else(|| false);
-
-    // TODO(2025-03-17 aubin) - Add generic validation on node upserts instead of duplicating it for folders, tables, documents.
-    if payload.parents.get(0) != Some(&payload.document_id) {
-        return error_response(
-            StatusCode::BAD_REQUEST,
-            "invalid_parents",
-            "Failed to upsert document - parents[0] and document_id should be equal",
-            None,
-        );
-    }
-
-    match &payload.parent_id {
-        Some(parent_id) => {
-            if payload.parents.get(1) != Some(parent_id) {
-                return error_response(
-                    StatusCode::BAD_REQUEST,
-                    "invalid_parent_id",
-                    "Failed to upsert document - parents[1] and parent_id should be equal",
-                    None,
-                );
-            }
-        }
-        None => {
-            if payload.parents.len() > 1 {
-                return error_response(
-                    StatusCode::BAD_REQUEST,
-                    "invalid_parent_id",
-                    "Failed to upsert document - parent_id should not be null if parents[1] is defined",
-                    None,
-                );
-            }
-        }
-    }
-
-    if payload.title.trim().is_empty() {
-        return error_response(
-            StatusCode::BAD_REQUEST,
-            "title_is_empty",
-            "Failed to upsert document - title is empty",
-            None,
-        );
-    }
-
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => {
-                match ds
-                    .upsert(
-                        payload.credentials,
-                        state.store.clone(),
-                        state.qdrant_clients.clone(),
-                        &payload.document_id,
-                        payload.title,
-                        payload.mime_type,
-                        &payload.provider_visibility,
-                        payload.timestamp,
-                        &payload.tags,
-                        &payload.parents,
-                        &payload.source_url,
-                        payload.section,
-                        true, // preserve system tags
-                        state.search_store.clone(),
-                    )
-                    .await
-                {
-                    Err(e) => error_response(
-                        StatusCode::INTERNAL_SERVER_ERROR,
-                        "internal_server_error",
-                        "Failed to upsert document",
-                        Some(e),
-                    ),
-                    Ok(d) => {
-                        let mut response_data = json!({
-                            "data_source": {
-                                "created": ds.created(),
-                                "data_source_id": ds.data_source_id(),
-                                "config": ds.config(),
-                            },
-                        });
-                        if light_document_output {
-                            response_data["document"] = json!({
-                                "hash": d.hash,
-                                "text_size": d.text_size,
-                                "chunk_count": d.chunk_count,
-                                "token_count": d.token_count,
-                                "created": d.created,
-                            });
-                        } else {
-                            response_data["document"] = json!(d);
-                        }
-                        (
-                            StatusCode::OK,
-                            Json(APIResponse {
-                                error: None,
-                                response: Some(response_data),
-                            }),
-                        )
-                    }
-                }
-            }
-        },
-    }
-}
-
-// Get a document blob from a data source.
-
-async fn data_sources_documents_retrieve_blob(
-    Path((project_id, data_source_id, document_id)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => match ds
-                .retrieve_api_blob(state.store.clone(), &document_id)
-                .await
-            {
-                Err(e) => error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to retrieve document blob",
-                    Some(e),
-                ),
-                Ok(None) => error_response(
-                    StatusCode::NOT_FOUND,
-                    "data_source_document_not_found",
-                    &format!("No document found for id `{}`", document_id),
-                    None,
-                ),
-                Ok(Some(blob)) => {
-                    let blob_value = serde_json::to_value(blob).unwrap();
-                    (
-                        StatusCode::OK,
-                        Json(APIResponse {
-                            error: None,
-                            response: Some(blob_value),
-                        }),
-                    )
-                }
-            },
-        },
-    }
-}
-
-/// List documents from a data source.
-
-#[derive(serde::Deserialize)]
-struct DataSourcesListQuery {
-    document_ids: Option<String>, // Parse as JSON.
-    limit: Option<usize>,
-    offset: Option<usize>,
-    view_filter: Option<String>, // Parsed as JSON.
-}
-
-async fn data_sources_documents_list(
-    Path((project_id, data_source_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-    Query(query): Query<DataSourcesListQuery>,
-) -> (StatusCode, Json<APIResponse>) {
-    let view_filter: Option<SearchFilter> = match query
-        .view_filter
-        .as_ref()
-        .and_then(|f| Some(serde_json::from_str(f)))
-    {
-        Some(Ok(f)) => Some(f),
-        None => None,
-        Some(Err(e)) => {
-            return error_response(
-                StatusCode::BAD_REQUEST,
-                "invalid_view_filter",
-                "Failed to parse view_filter query parameter",
-                Some(e.into()),
-            )
-        }
-    };
-
-    let limit_offset: Option<(usize, usize)> = match (query.limit, query.offset) {
-        (Some(limit), Some(offset)) => Some((limit, offset)),
-        _ => None,
-    };
-
-    let document_ids: Option<Vec<String>> = match query.document_ids {
-        Some(ref ids) => match serde_json::from_str(ids) {
-            Ok(parsed_ids) => Some(parsed_ids),
-            Err(e) => {
-                return error_response(
-                    StatusCode::BAD_REQUEST,
-                    "invalid_document_ids",
-                    "Failed to parse document_ids query parameter",
-                    Some(e.into()),
-                )
-            }
-        },
-        None => None,
-    };
-
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .list_data_source_documents(
-            &project,
-            &data_source_id,
-            &match view_filter {
-                Some(filter) => Some(filter.postprocess_for_data_source(&data_source_id)),
-                None => None,
-            },
-            &document_ids,
-            limit_offset,
-            true, // remove system tags
-        )
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to list data source",
-            Some(e),
-        ),
-        Ok(documents) => {
-            let stats = state
-                .search_store
-                .get_data_source_stats(vec![data_source_id.clone()])
-                .await;
-
-            let total = match stats {
-                Ok((data_source_stats, _)) => data_source_stats
-                    .first()
-                    .map(|ds| ds.document_count as usize)
-                    .unwrap_or(0),
-                Err(e) => {
-                    error!(
-                        error = %e,
-                        data_source_id = data_source_id,
-                        "Failed to get document count from Elasticsearch"
-                    );
-                    0
-                }
-            };
-
-            return (
-                StatusCode::OK,
-                Json(APIResponse {
-                    error: None,
-                    response: Some(json!({
-                        "documents": documents,
-                        "limit": query.limit,
-                        "offset": query.offset,
-                        "total": total,
-                    })),
-                }),
-            );
-        }
-    }
-}
-
-/// Retrieve document from a data source.
-#[derive(serde::Deserialize)]
-struct DataSourcesDocumentsRetrieveQuery {
-    version_hash: Option<String>,
-    view_filter: Option<String>, // Parsed as JSON.
-}
-
-async fn data_sources_documents_retrieve(
-    Path((project_id, data_source_id, document_id)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-    Query(query): Query<DataSourcesDocumentsRetrieveQuery>,
-) -> (StatusCode, Json<APIResponse>) {
-    let view_filter: Option<SearchFilter> = match query
-        .view_filter
-        .as_ref()
-        .and_then(|f| Some(serde_json::from_str(f)))
-    {
-        Some(Ok(f)) => Some(f),
-        None => None,
-        Some(Err(e)) => {
-            return error_response(
-                StatusCode::BAD_REQUEST,
-                "invalid_view_filter",
-                "Failed to parse view_filter query parameter",
-                Some(e.into()),
-            )
-        }
-    };
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => match ds
-                .retrieve(
-                    state.store.clone(),
-                    &document_id,
-                    &match view_filter {
-                        Some(filter) => Some(filter.postprocess_for_data_source(&data_source_id)),
-                        None => None,
-                    },
-                    true,
-                    &query.version_hash,
-                )
-                .await
-            {
-                Err(e) => error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to retrieve document",
-                    Some(e),
-                ),
-                Ok(None) => error_response(
-                    StatusCode::NOT_FOUND,
-                    "data_source_document_not_found",
-                    &format!("No document found for id `{}`", document_id),
-                    None,
-                ),
-                Ok(Some(d)) => (
-                    StatusCode::OK,
-                    Json(APIResponse {
-                        error: None,
-                        response: Some(json!({
-                            "document": d,
-                            "data_source": {
-                                "created": ds.created(),
-                                "data_source_id": ds.data_source_id(),
-                                "config": ds.config(),
-                                "name": ds.name(),
-                            },
-                        })),
-                    }),
-                ),
-            },
-        },
-    }
-}
-
-/// Retrieve document text from a data source with offset and limit.
-#[derive(serde::Deserialize)]
-struct DataSourcesDocumentsRetrieveTextQuery {
-    offset: Option<usize>,
-    limit: Option<usize>,
-    grep: Option<String>,
-    version_hash: Option<String>,
-    view_filter: Option<String>, // Parsed as JSON.
-}
-
-async fn data_sources_documents_retrieve_text(
-    Path((project_id, data_source_id, document_id)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-    Query(query): Query<DataSourcesDocumentsRetrieveTextQuery>,
-) -> (StatusCode, Json<APIResponse>) {
-    // Call the existing retrieve function
-    let retrieve_query = DataSourcesDocumentsRetrieveQuery {
-        version_hash: query.version_hash,
-        view_filter: query.view_filter,
-    };
-
-    let (status, json_response) = data_sources_documents_retrieve(
-        Path((project_id, data_source_id, document_id)),
-        State(state),
-        Query(retrieve_query),
-    )
-    .await;
-
-    // If the request failed, return the error as-is
-    if status != StatusCode::OK {
-        return (status, json_response);
-    }
-
-    // Extract the document text from the response
-    let text = json_response
-        .response
-        .as_ref()
-        .and_then(|r| r.get("document"))
-        .and_then(|d| d.get("text"))
-        .and_then(|t| t.as_str());
-
-    let text = match text {
-        Some(t) => t,
-        None => {
-            return error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "internal_server_error",
-                "Failed to extract text from document response",
-                None,
-            )
-        }
-    };
-
-    // First apply character-based offset and limit
-    let offset = query.offset.unwrap_or(0);
-    let limit = query.limit;
-
-    let text_len = text.len();
-    let start = offset.min(text_len);
-    let end = match limit {
-        Some(l) => (start + l).min(text_len),
-        None => text_len,
-    };
-
-    let text_slice = &text[start..end];
-
-    // Then apply grep filter if provided
-    let filtered_text = match &query.grep {
-        Some(pattern) => match Regex::new(pattern) {
-            Ok(re) => {
-                let lines: Vec<&str> = text_slice
-                    .lines()
-                    .filter(|line| re.is_match(line))
-                    .collect();
-                lines.join("\n")
-            }
-            Err(_) => {
-                return error_response(
-                    StatusCode::BAD_REQUEST,
-                    "invalid_regex",
-                    &format!("Invalid regular expression: {}", pattern),
-                    None,
-                )
-            }
-        },
-        None => text_slice.to_string(),
-    };
-
-    (
-        StatusCode::OK,
-        Json(APIResponse {
-            error: None,
-            response: Some(json!({
-                "text": filtered_text,
-                "total_characters": text_len,
-                "offset": start,
-                "limit": limit,
-            })),
-        }),
-    )
-}
-
-/// Delete document from a data source.
-
-async fn data_sources_documents_delete(
-    Path((project_id, data_source_id, document_id)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => match ds
-                .delete_document(
-                    state.store.clone(),
-                    state.qdrant_clients.clone(),
-                    state.search_store.clone(),
-                    &document_id,
-                )
-                .await
-            {
-                Err(e) => error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to delete document",
-                    Some(e),
-                ),
-                Ok(_) => (
-                    StatusCode::OK,
-                    Json(APIResponse {
-                        error: None,
-                        response: Some(json!({
-                            "data_source": {
-                                "created": ds.created(),
-                                "data_source_id": ds.data_source_id(),
-                                "config": ds.config(),
-                            },
-                        })),
-                    }),
-                ),
-            },
-        },
-    }
-}
-
-/// Scrub document deleted versions
-
-async fn data_sources_documents_scrub_deleted_versions(
-    Path((project_id, data_source_id, document_id)): Path<(i64, String, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => match ds
-                .scrub_document_deleted_versions(state.store.clone(), &document_id)
-                .await
-            {
-                Err(e) => error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to scrub document deleted versions",
-                    Some(e),
-                ),
-                Ok(versions) => (
-                    StatusCode::OK,
-                    Json(APIResponse {
-                        error: None,
-                        response: Some(json!({
-                            "versions": versions,
-                        })),
-                    }),
-                ),
-            },
-        },
-    }
-}
-
-/// Delete a data source.
-
-async fn data_sources_delete(
-    Path((project_id, data_source_id)): Path<(i64, String)>,
-    State(state): State<Arc<APIState>>,
-) -> (StatusCode, Json<APIResponse>) {
-    let project = project::Project::new_from_id(project_id);
-    match state
-        .store
-        .load_data_source(&project, &data_source_id)
-        .await
-    {
-        Err(e) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "internal_server_error",
-            "Failed to retrieve data source",
-            Some(e),
-        ),
-        Ok(ds) => match ds {
-            None => error_response(
-                StatusCode::NOT_FOUND,
-                "data_source_not_found",
-                &format!("No data source found for id `{}`", data_source_id),
-                None,
-            ),
-            Some(ds) => match ds
-                .delete(
-                    state.store.clone(),
-                    state.databases_store.clone(),
-                    state.qdrant_clients.clone(),
-                    state.search_store.clone(),
-                )
-                .await
-            {
-                Err(e) => error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "internal_server_error",
-                    "Failed to delete data source",
-                    Some(e),
-                ),
-                Ok(_) => (
-                    StatusCode::OK,
-                    Json(APIResponse {
-                        error: None,
-                        response: Some(json!({
-                            "data_source": {
-                                "created": ds.created(),
-                                "data_source_id": ds.data_source_id(),
-                                "config": ds.config(),
-                            }
-                        })),
-                    }),
-                ),
-            },
-        },
-    }
 }
 
 #[derive(serde::Deserialize)]
@@ -3992,121 +1739,121 @@ fn main() {
         // Specifications
         .route(
             "/projects/{project_id}/specifications/check",
-            post(specifications_check),
+            post(specifications::specifications_check),
         )
         .route(
             "/projects/{project_id}/specifications/{hash}",
-            get(specifications_retrieve),
+            get(specifications::specifications_retrieve),
         )
         .route(
             "/projects/{project_id}/specifications",
-            get(specifications_get),
+            get(specifications::specifications_get),
         )
         .route(
             "/projects/{project_id}/specifications",
-            post(specifications_post),
+            post(specifications::specifications_post),
         )
 
         // Datasets
-        .route("/projects/{project_id}/datasets", post(datasets_register))
-        .route("/projects/{project_id}/datasets", get(datasets_list))
+        .route("/projects/{project_id}/datasets", post(datasets::datasets_register))
+        .route("/projects/{project_id}/datasets", get(datasets::datasets_list))
         .route(
             "/projects/{project_id}/datasets/{dataset_id}/{hash}",
-            get(datasets_retrieve),
+            get(datasets::datasets_retrieve),
         )
         // Runs
-        .route("/projects/{project_id}/runs", post(runs_create))
+        .route("/projects/{project_id}/runs", post(runs::runs_create))
         .route(
             "/projects/{project_id}/runs/stream",
-            post(runs_create_stream),
+            post(runs::runs_create_stream),
         )
-        .route("/projects/{project_id}/runs", get(runs_list))
+        .route("/projects/{project_id}/runs", get(runs::runs_list))
         .route(
             "/projects/{project_id}/runs/batch",
-            post(runs_retrieve_batch),
+            post(runs::runs_retrieve_batch),
         )
-        .route("/projects/{project_id}/runs/{run_id}", get(runs_retrieve))
+        .route("/projects/{project_id}/runs/{run_id}", get(runs::runs_retrieve))
         .route(
             "/projects/{project_id}/runs/{run_id}",
-            delete(runs_delete),
+            delete(runs::runs_delete),
         )
         .route(
             "/projects/{project_id}/runs/{run_id}/cancel",
-            post(runs_cancel),
+            post(runs::runs_cancel),
         )
         .route(
             "/projects/{project_id}/runs/{run_id}/blocks/{block_type}/{block_name}",
-            get(runs_retrieve_block),
+            get(runs::runs_retrieve_block),
         )
         .route(
             "/projects/{project_id}/runs/{run_id}/status",
-            get(runs_retrieve_status),
+            get(runs::runs_retrieve_status),
         )
         // DataSources
         .route(
             "/projects/{project_id}/data_sources",
-            post(data_sources_register),
+            post(data_sources::data_sources_register),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}",
-            patch(data_sources_update),
+            patch(data_sources::data_sources_update),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}",
-            get(data_sources_retrieve),
+            get(data_sources::data_sources_retrieve),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/tokenize",
-            post(data_sources_tokenize),
+            post(data_sources::data_sources_tokenize),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents/{document_id}/versions",
-            get(data_sources_documents_versions_list),
+            get(data_sources::data_sources_documents_versions_list),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents",
-            post(data_sources_documents_upsert),
+            post(data_sources::data_sources_documents_upsert),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents/{document_id}/blob",
-            get(data_sources_documents_retrieve_blob),
+            get(data_sources::data_sources_documents_retrieve_blob),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents/{document_id}/tags",
-            patch(data_sources_documents_update_tags),
+            patch(data_sources::data_sources_documents_update_tags),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents/{document_id}/parents",
-            patch(data_sources_documents_update_parents),
+            patch(data_sources::data_sources_documents_update_parents),
         )
         // Provided by the data_source block.
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/search",
-            post(data_sources_search),
+            post(data_sources::data_sources_search),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents",
-            get(data_sources_documents_list),
+            get(data_sources::data_sources_documents_list),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents/{document_id}",
-            get(data_sources_documents_retrieve),
+            get(data_sources::data_sources_documents_retrieve),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents/{document_id}/text",
-            get(data_sources_documents_retrieve_text),
+            get(data_sources::data_sources_documents_retrieve_text),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents/{document_id}",
-            delete(data_sources_documents_delete),
+            delete(data_sources::data_sources_documents_delete),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}/documents/{document_id}/scrub_deleted_versions",
-            post(data_sources_documents_scrub_deleted_versions),
+            post(data_sources::data_sources_documents_scrub_deleted_versions),
         )
         .route(
             "/projects/{project_id}/data_sources/{data_source_id}",
-            delete(data_sources_delete),
+            delete(data_sources::data_sources_delete),
         )
         // Databases
         .route(
