+++ core/src/lib.rs
@@ -176,6 +176,10 @@ pub mod mem_check;
 
 pub mod api {
     pub mod api_state;
+    pub mod data_sources;
+    pub mod datasets;
     pub mod projects;
     pub(crate) mod run_manager;
+    pub mod runs;
+    pub mod specifications;
 }
